payment-handler/payment-request-event-untrusted.https.html

<!DOCTYPE html>
<meta charset=utf-8>
<link rel="help" href="https://w3c.github.io/payment-handler/">
<title>Test for untrusted payment request event</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>

async function getResultFromSW(serviceWorkerContainer) {
  return new Promise((resolve, reject) => {
    serviceWorkerContainer.addEventListener('message', listener = e => {
      serviceWorkerContainer.removeEventListener('message', listener);
      if (e.data) {
        resolve(e.data);
      } else {
        reject();
      }
    });
  });
}

promise_test(async t => {
  const registration = await service_worker_unregister_and_register(
      t, 'payment-request-event-untrusted.js', 'payment-app/');
  await wait_for_state(t, registration.installing, 'activated');

  const controlled_window = (await with_iframe('payment-app/payment.html')).contentWindow;
  const result = getResultFromSW(controlled_window.navigator.serviceWorker);
  controlled_window.navigator.serviceWorker.controller.postMessage('start');

  const expected = [
    "InvalidStateError",
    "InvalidStateError"
  ];

  assert_array_equals(await result, expected);
});

</script>
